pyproject.toml
[project]
name = "genai-graphrag-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "certifi==2025.10.5",
    "colorama==0.4.6",
    "distro==1.9.0",
    "fsspec==2024.12.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "jiter==0.12.0",
    "json-repair==0.44.1",
    "langchain-text-splitters>=0.3.0",
    "neo4j==5.28.2",
    "neo4j-graphrag==1.10.1",
    "numpy==2.3.4",
    "openai==2.7.1",
    "pydantic==2.12.4",
    "pydantic-core==2.41.5",
    "pypdf==6.2.0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "scipy==1.16.3",
    "sniffio==1.3.1",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "types-pyyaml==6.0.12.20250915",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
target-version = "py314"
line-length = 88
indent-width = 4

# Exclude common directories
exclude = [
    ".venv",
    "venv",
    "__pycache__",
    ".git",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # Pylint
    "PERF",   # Perflint
    "D",      # pydocstyle (docstrings)
    "ANN",    # flake8-annotations (type hints)
]
ignore = [
    "E501",    # line too long (handled by formatter)
    "E402",    # module level import not at top (load_dotenv() pattern is intentional)
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "ANN401",  # dynamically typed expressions (Any) are disallowed
    "D100",    # missing docstring in public module
    "D104",    # missing docstring in public package
]

[tool.ruff.lint.isort]
known-first-party = ["genai_graphrag_python"]
force-single-line = false
combine-as-imports = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
suppress-none-returning = true

[tool.ruff.lint.per-file-ignores]
"genai-graphrag-python/examples/*" = ["ERA001", "T201", "D", "ANN", "TC", "PTH"]  # Examples: relaxed for educational code
"genai-graphrag-python/solutions/*" = ["ERA001", "T201", "D", "ANN", "TC", "PTH"]  # Solutions: relaxed for educational code
"genai-graphrag-python/*.py" = ["T201", "D103", "ANN", "PTH", "SIM115"]  # Main scripts: allow prints, os.path patterns
"**/test_*.py" = ["D", "ANN", "PLC0415", "SIM"]  # Tests: relaxed for test patterns
"main.py" = ["D103"]  # Main entry point: no docstring required

[tool.ruff.lint.pylint]
max-args = 8
